+++ src/features/island/hud/Hud.tsx
@@ -21,7 +21,7 @@ import { useLocation } from "react-router-dom";
 import { useIsMobile } from "lib/utils/hooks/useIsMobile";
 import { createPortal } from "react-dom";
 import { HalveningCountdown } from "./components/HalveningCountdown";
-import { Deliveries } from "./components/deliveries/Deliveries";
+import { DeliveryButton } from "./components/deliveries/DeliveryButton";
 import { hasFeatureAccess } from "lib/flags";
 
 /**
@@ -53,14 +53,11 @@ const HudComponent: React.FC<{
     gameService.send("DEPOSIT", args);
   };
 
-  const landId = gameState.context.state.id;
-
   const user = authService.state.context.user;
   const isFullUser = user.type === "FULL";
   const farmAddress = isFullUser ? user.farmAddress : undefined;
 
   const isDawnBreakerIsland = location.pathname.includes("dawn-breaker");
-  const isHelios = location.pathname.includes("helios");
 
   return (
     <>
@@ -145,7 +142,7 @@ const HudComponent: React.FC<{
             "NEW_DELIVERIES"
           ) && (
             <div className="fixed z-50 bottom-0 left-0">
-              <Deliveries />
+              <DeliveryButton />
             </div>
           )}
           <HalveningCountdown />
